Pulled out some common camera code
